pyproject.toml
[project]
name = "vibecoded-tictactoe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "python-dotenv>=1.2.1",
    "httpx>=0.27.2",
]
